add breaking test for `upsert`
There is a quirk in this mixin: if you `upsert` a model (i.e. you just know the id and properties, but you don't want to `find` it first) the createdAt-Timestamp is updated, too. This must not happen, of course.
I fear it's related to the lb-ds-juggler's implementation of default values.

Can you think of a solution besides querying for the model everytime?