ES6
gulp + babel + source maps

Source Maps weren’t as helpful as one would hope, it seemed to confuse
column numbers and line numbers.  Most errors were reported on line 9
with a column number in the two hundreds.

The es6 code is clean but overall very similar.

Object.assign requires shims in babel so we needed to pull in a plugin
to help convert that function.