new test fixtures and working debug
The old test is now test-index which tests the deprecated index file
method.
New tests are using a loopback app in the fixtures directory to
simulate loading the mixin.
Using require for the debug module, unlikely it more useful than just
requiring it without the local module as of yet but may want to try
again later.